requirements.txt
mss>=9.0.1
Pillow>=10.0.0
requests>=2.31.0
keyboard>=0.13.5
python-dotenv>=1.0.0
psutil>=5.9.0

# OCR
pytesseract>=0.3.10

# Server dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
SQLAlchemy>=2.0.25
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0

# Async support
aiofiles>=23.2.1

# Gemini API (add if not already installed)
gemini-webapi>=2.0.0